Form API and handling

+++ docker/gunicorn.sh
@@ -3,14 +3,23 @@
 set -o errexit
 set -o nounset
 
-echo 'Copying static assets...'
-rm -rf /dist/*
-cp -r /code/dist/* /dist/
-
 # Start gunicorn with workers:
-/usr/local/bin/gunicorn main:app \
-  --workers 2 \
-  --bind 0.0.0.0:8080 \
-  --chdir=/code/app \
-  --log-file=- \
-  --worker-tmp-dir /dev/shm
+if [ "$FLASK_ENV" = "deploy" ]
+then
+  echo 'Copying static assets...'
+  rm -rf /dist/*
+  cp -r /code/dist/* /dist/
+  /usr/local/bin/gunicorn main:app \
+    --workers 4 \
+    --bind 0.0.0.0:8080 \
+    --chdir=/code/app \
+    --log-file=- \
+    --worker-tmp-dir /dev/shm
+else
+  /usr/local/bin/gunicorn main:app --reload \
+    --workers 2 \
+    --bind 0.0.0.0:8080 \
+    --chdir=/code/app \
+    --log-file=- \
+    --worker-tmp-dir /dev/shm
+fi
